+++ apps/trpc-app/src/server/routes/trpc/[trpc].ts
@@ -1,6 +1,6 @@
 import { appRouter } from '../../trpc/routers';
 import { createContext } from '../../trpc/context';
-import { createTrpcNitroHandler } from '@analogjs/trpc';
+import { createTrpcNitroHandler } from '@analogjs/trpc/server';
 // export API handler
 export default createTrpcNitroHandler({
   router: appRouter,
