+++ packages/trpc/src/index.ts
@@ -1,5 +1,4 @@
 export * from './lib/client/client';
-export * from './lib/server/server';
 export * from './lib/utils/wait-for';
 
 export { CreateTrpcProxyClient } from './lib/client/trpc-rxjs-proxy';
